# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: South African School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      summary: List fee records
      parameters:
        - name: learnerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeeRecord"
    post:
      operationId: createFeeRecord
      tags: [fees]
      summary: Create fee record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeeRecordBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
  /fees/{id}:
    patch:
      operationId: updateFeeRecord
      tags: [fees]
      summary: Update fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeeRecordBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeRecord"
    delete:
      operationId: deleteFeeRecord
      tags: [fees]
      summary: Delete fee record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /dashboard/teacher/{userId}:
    get:
      operationId: getTeacherDashboard
      tags: [dashboard]
      summary: Get teacher dashboard summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teacher dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherDashboard"
  /dashboard/learner/{userId}:
    get:
      operationId: getLearnerDashboard
      tags: [dashboard]
      summary: Get learner dashboard summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Learner dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerDashboard"
  /dashboard/parent/{userId}:
    get:
      operationId: getParentDashboard
      tags: [dashboard]
      summary: Get parent dashboard summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parent dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParentDashboard"
...
components:
  schemas:
...
    Learner:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        admissionNumber:
          type: string
        fullName:
          type: string
        grade:
          type: string
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        parentId:
          type: ["integer", "null"]
        parentName:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - admissionNumber
        - fullName
        - grade
        - createdAt
...
    Class:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        grade:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        learnerCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - grade
        - learnerCount
        - createdAt
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        grade:
          type: string
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - grade
        - createdAt
...
    Mark:
      type: object
      properties:
        id:
          type: integer
        learnerId:
          type: integer
        learnerName:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        assessmentType:
          type: string
          enum: [test, assignment, exam, project, classwork]
        assessmentName:
          type: string
        score:
          type: number
        maxScore:
          type: number
        percentage:
          type: number
        term:
          type: integer
        grade:
          type: ["string", "null"]
        recordedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - learnerId
        - subjectId
        - assessmentType
        - assessmentName
        - score
        - maxScore
        - percentage
        - term
        - createdAt
...
    LearnerProgress:
      type: object
      properties:
        learnerId:
          type: integer
        learnerName:
          type: string
        grade:
          type: string
        overallAverage:
          type: number
        subjectAverages:
          type: array
          items:
            $ref: "#/components/schemas/SubjectAverage"
        recentMarks:
          type: array
          items:
            $ref: "#/components/schemas/Mark"
      required:
        - learnerId
        - learnerName
        - grade
        - overallAverage
        - subjectAverages
        - recentMarks
    SubjectAverage:
      type: object
      properties:
        subjectId:
          type: integer
        subjectName:
          type: string
        average:
          type: number
        term:
          type: integer
      required:
        - subjectId
        - subjectName
        - average
        - term
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [general, academic, financial, attendance, urgent]
        targetRole:
          type: ["string", "null"]
        targetUserId:
          type: ["integer", "null"]
        isRead:
          type: boolean
        sentBy:
          type: ["integer", "null"]
        senderName:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - message
        - type
        - isRead
        - createdAt
...
    FeeRecord:
      type: object
      properties:
        id:
          type: integer
        learnerId:
          type: integer
        learnerName:
          type: ["string", "null"]
        description:
          type: string
        totalAmount:
          type: number
        amountPaid:
          type: number
        outstandingBalance:
          type: number
        dueDate:
          type: string
        status:
          type: string
          enum: [outstanding, paid, partial, overdue]
        term:
          type: integer
        year:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - learnerId
        - description
        - totalAmount
        - amountPaid
        - outstandingBalance
        - dueDate
        - status
        - term
        - year
        - createdAt
    CreateFeeRecordBody:
      type: object
      properties:
        learnerId:
          type: integer
        description:
          type: string
        totalAmount:
          type: number
        dueDate:
          type: string
        term:
          type: integer
        year:
          type: integer
      required:
        - learnerId
        - description
        - totalAmount
        - dueDate
        - term
        - year
    UpdateFeeRecordBody:
      type: object
      properties:
        description:
          type: ["string", "null"]
        totalAmount:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        feeId:
          type: integer
        learnerId:
          type: integer
        learnerName:
          type: ["string", "null"]
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [eft, credit_card, debit_card, cash, direct_deposit]
        referenceNumber:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        processedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - feeId
        - learnerId
        - amount
        - paymentMethod
        - referenceNumber
        - status
        - createdAt
...
    TimetableEntry:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
        teacherName:
          type: ["string", "null"]
        dayOfWeek:
          type: string
          enum: [Monday, Tuesday, Wednesday, Thursday, Friday]
        startTime:
          type: string
        endTime:
          type: string
        room:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - classId
        - subjectId
        - dayOfWeek
        - startTime
        - endTime
        - createdAt
...
    TeacherDashboard:
      type: object
      properties:
        myClasses:
          type: array
          items:
            $ref: "#/components/schemas/Class"
        mySubjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
        totalLearners:
          type: integer
        attendanceTodayPct:
          type: number
        recentNotifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        pendingAssessments:
          type: integer
      required:
        - myClasses
        - mySubjects
        - totalLearners
        - attendanceTodayPct
        - recentNotifications
        - pendingAssessments
    LearnerDashboard:
      type: object
      properties:
        learner:
          $ref: "#/components/schemas/Learner"
        overallAverage:
          type: number
        attendancePct:
          type: number
        recentMarks:
          type: array
          items:
            $ref: "#/components/schemas/Mark"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        timetableToday:
          type: array
          items:
            $ref: "#/components/schemas/TimetableEntry"
      required:
        - overallAverage
        - attendancePct
        - recentMarks
        - notifications
        - timetableToday
    ParentDashboard:
      type: object
      properties:
        children:
          type: array
          items:
            $ref: "#/components/schemas/Learner"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        outstandingFees:
          type: number
        recentPayments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        childrenProgress:
          type: array
          items:
            $ref: "#/components/schemas/LearnerProgress"
      required:
        - children
        - notifications
        - outstandingFees
        - recentPayments
        - childrenProgress
...